+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change log for SharePointDsc
 
-## Unreleased
+## UNRELEASED
+
+* SPFarmSolution
+  * Fix for Web Application scoped solutions.
 
 * SPWebAppAuthentication
   * Fixes issue where Test method return false on NON-US OS.